+++ my-app/pages/_app.js
@@ -7,6 +7,5 @@ export default function App({ Component, pageProps }) {
   <>
 
       <Component {...pageProps} />
-      <Footer />
   </>
   )}
